refactor: rename `plugin-inputs.ts` to `plugin-input.ts`
